# backend/api/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "课程管理系统API",
    "description": "课程管理系统后端API文档",
    "version": "1.0.0",
    "contact": {
      "name": "技术支持",
      "email": "support@example.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "开发环境"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "integer",
            "description": "用户ID"
          },
          "username": {
            "type": "string",
            "description": "用户名"
          },
          "name": {
            "type": "string",
            "description": "真实姓名"
          },
          "roleId": {
            "type": "integer",
            "description": "角色ID"
          },
          "roleName": {
            "type": "string",
            "description": "角色名称"
          },
          "email": {
            "type": "string",
            "description": "邮箱"
          },
          "phone": {
            "type": "string",
            "description": "手机号"
          },
          "avatar": {
            "type": "string",
            "description": "头像URL"
          },
          "departmentId": {
            "type": "integer",
            "description": "部门ID"
          },
          "departmentName": {
            "type": "string",
            "description": "部门名称"
          },
          "position": {
            "type": "string",
            "description": "职位"
          },
          "status": {
            "type": "integer",
            "description": "状态：1启用，0禁用"
          },
          "joinDate": {
            "type": "string",
            "format": "date-time",
            "description": "入职日期"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "创建时间"
          }
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": ["username", "password"],
        "properties": {
          "username": {
            "type": "string",
            "description": "用户名"
          },
          "password": {
            "type": "string",
            "description": "密码"
          }
        }
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "操作是否成功"
          },
          "data": {
            "type": "object",
            "properties": {
              "user": {
                "$ref": "#/components/schemas/User"
              },
              "token": {
                "type": "string",
                "description": "认证令牌"
              }
            }
          },
          "message": {
            "type": "string",
            "description": "操作消息"
          }
        }
      },
      "Schedule": {
        "type": "object",
        "properties": {
          "scheduleId": {
            "type": "integer",
            "description": "排课ID"
          },
          "courseId": {
            "type": "integer",
            "description": "课程ID"
          },
          "courseName": {
            "type": "string",
            "description": "课程名称"
          },
          "teacherId": {
            "type": "integer",
            "description": "教师ID"
          },
          "teacherName": {
            "type": "string",
            "description": "教师姓名"
          },
          "classroomId": {
            "type": "integer",
            "description": "教室ID"
          },
          "classroomName": {
            "type": "string",
            "description": "教室名称"
          },
          "startTime": {
            "type": "string",
            "format": "date-time",
            "description": "开始时间"
          },
          "endTime": {
            "type": "string",
            "format": "date-time",
            "description": "结束时间"
          },
          "duration": {
            "type": "integer",
            "description": "课时时长（分钟）"
          },
          "status": {
            "type": "string",
            "description": "状态：scheduled(已排课), cancelled(已取消), completed(已完成), in_progress(进行中)"
          },
          "remark": {
            "type": "string",
            "description": "备注"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "创建时间"
          }
        }
      },
      "Attendance": {
        "type": "object",
        "properties": {
          "attendanceId": {
            "type": "integer",
            "description": "签到ID"
          },
          "scheduleId": {
            "type": "integer",
            "description": "排课ID"
          },
          "studentId": {
            "type": "integer",
            "description": "学生ID"
          },
          "studentName": {
            "type": "string",
            "description": "学生姓名"
          },
          "attendanceTime": {
            "type": "string",
            "format": "date-time",
            "description": "签到时间"
          },
          "leaveTime": {
            "type": "string",
            "format": "date-time",
            "description": "签退时间"
          },
          "status": {
            "type": "string",
            "description": "状态：present(正常), late(迟到), early_leave(早退), late_early_leave(迟到早退), absent(缺勤)"
          },
          "lateMinutes": {
            "type": "integer",
            "description": "迟到分钟数"
          },
          "earlyLeaveMinutes": {
            "type": "integer",
            "description": "早退分钟数"
          },
          "attendanceType": {
            "type": "string",
            "description": "签到方式"
          },
          "leaveType": {
            "type": "string",
            "description": "签退方式"
          }
        }
      },
      "Record": {
        "type": "object",
        "properties": {
          "recordId": {
            "type": "integer",
            "description": "上课记录ID"
          },
          "scheduleId": {
            "type": "integer",
            "description": "排课ID"
          },
          "teacherId": {
            "type": "integer",
            "description": "教师ID"
          },
          "teacherName": {
            "type": "string",
            "description": "教师姓名"
          },
          "courseId": {
            "type": "integer",
            "description": "课程ID"
          },
          "courseName": {
            "type": "string",
            "description": "课程名称"
          },
          "actualStartTime": {
            "type": "string",
            "format": "date-time",
            "description": "实际开始时间"
          },
          "actualEndTime": {
            "type": "string",
            "format": "date-time",
            "description": "实际结束时间"
          },
          "content": {
            "type": "string",
            "description": "上课内容"
          },
          "homework": {
            "type": "string",
            "description": "作业"
          },
          "presentCount": {
            "type": "integer",
            "description": "出勤人数"
          },
          "lateCount": {
            "type": "integer",
            "description": "迟到人数"
          },
          "absentCount": {
            "type": "integer",
            "description": "缺勤人数"
          },
          "status": {
            "type": "string",
            "description": "状态"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "创建时间"
          }
        }
      },
      "Deduction": {
        "type": "object",
        "properties": {
          "deductionId": {
            "type": "integer",
            "description": "扣课记录ID"
          },
          "studentId": {
            "type": "integer",
            "description": "学生ID"
          },
          "studentName": {
            "type": "string",
            "description": "学生姓名"
          },
          "courseId": {
            "type": "integer",
            "description": "课程ID"
          },
          "courseName": {
            "type": "string",
            "description": "课程名称"
          },
          "deductionAmount": {
            "type": "number",
            "description": "扣课数量"
          },
          "deductionType": {
            "type": "string",
            "description": "扣课类型: class_attendance(上课扣课), make_up(补课), adjustment(课时调整)"
          },
          "deductionTime": {
            "type": "string",
            "format": "date-time",
            "description": "扣课时间"
          },
          "operatorId": {
            "type": "integer",
            "description": "操作人ID"
          },
          "operatorName": {
            "type": "string",
            "description": "操作人姓名"
          },
          "remark": {
            "type": "string",
            "description": "备注"
          }
        }
      },
      "PaginationResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "操作是否成功"
          },
          "data": {
            "type": "object",
            "properties": {
              "list": {
                "type": "array",
                "items": {
                  "type": "object"
                },
                "description": "数据列表"
              },
              "total": {
                "type": "integer",
                "description": "总记录数"
              },
              "page": {
                "type": "integer",
                "description": "当前页码"
              },
              "pageSize": {
                "type": "integer",
                "description": "每页记录数"
              }
            }
          }
        }
      },
      "BaseResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "操作是否成功"
          },
          "message": {
            "type": "string",
            "description": "操作消息"
          },
          "data": {
            "type": "object",
            "description": "响应数据"
          }
        }
      }
    }
  },
  "paths": {
    "/auth/login": {
      "post": {
        "summary": "用户登录",
        "description": "用户登录接口，返回认证令牌和用户信息",
        "tags": ["用户认证"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "登录成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "401": {
            "description": "认证失败"
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "summary": "用户登出",
        "description": "用户登出接口",
        "tags": ["用户认证"],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "登出成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse"
                }
              }
            }
          }
        }
      }
    },
    "/users/profile": {
      "get": {
        "summary": "获取当前用户信息",
        "description": "获取当前登录用户的详细信息",
        "tags": ["用户管理"],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "更新用户信息",
        "description": "更新当前用户的个人信息",
        "tags": ["用户管理"],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "真实姓名"
                  },
                  "email": {
                    "type": "string",
                    "description": "邮箱"
                  },
                  "phone": {
                    "type": "string",
                    "description": "手机号"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse"
                }
              }
            }
          }
        }
      }
    },
    "/schedules": {
      "get": {
        "summary": "获取排课列表",
        "description": "获取排课列表，支持分页和筛选",
        "tags": ["排课管理"],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "页码",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "每页记录数",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "courseId",
            "in": "query",
            "description": "课程ID",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "teacherId",
            "in": "query",
            "description": "教师ID",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "startDate",
            "in": "query",
            "description": "开始日期",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "endDate",
            "in": "query",
            "description": "结束日期",
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginationResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "创建排课",
        "description": "创建新的排课记录",
        "tags": ["排课管理"],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "courseId": {
                    "type": "integer",
                    "description": "课程ID"
                  },
                  "teacherId": {
                    "type": "integer",
                    "description": "教师ID"
                  },
                  "classroomId": {
                    "type": "integer",
                    "description": "教室ID"
                  },
                  "startTime": {
                    "type": "string",
                    "format": "date-time",
                    "description": "开始时间"
                  },
                  "endTime": {
                    "type": "string",
                    "format": "date-time",
                    "description": "结束时间"
                  },
                  "remark": {
                    "type": "string",
                    "description": "备注"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse"
                }
              }
            }
          }
        }
      }
    },
    "/attendance/check-in": {
      "post": {
        "summary": "学生签到",
        "description": "学生签到接口",
        "tags": ["签到管理"],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "scheduleId": {
                    "type": "integer",
                    "description": "排课ID"
                  },
                  "studentId": {
                    "type": "integer",
                    "description": "学生ID"
                  },
                  "location": {
                    "type": "string",
                    "description": "签到位置"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "签到成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse"
                }
              }
            }
          }
        }
      }
    },
    "/attendance/check-out": {
      "post": {
        "summary": "学生签退",
        "description": "学生签退接口",
        "tags": ["签到管理"],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "scheduleId": {
                    "type": "integer",
                    "description": "排课ID"
                  },
                  "studentId": {
                    "type": "integer",
                    "description": "学生ID"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "签退成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse"
                }
              }
            }
          }
        }
      }
    },
    "/records": {
      "get": {
        "summary": "获取上课记录列表",
        "description": "获取上课记录列表，支持分页和筛选",
        "tags": ["上课记录"],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "页码",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "每页记录数",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "scheduleId",
            "in": "query",
            "description": "排课ID",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "teacherId",
            "in": "query",
            "description": "教师ID",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginationResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "创建上课记录",
        "description": "创建新的上课记录",
        "tags": ["上课记录"],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "scheduleId": {
                    "type": "integer",
                    "description": "排课ID"
                  },
                  "actualStartTime": {
                    "type": "string",
                    "format": "date-time",
                    "description": "实际开始时间"
                  },
                  "actualEndTime": {
                    "type": "string",
                    "format": "date-time",
                    "description": "实际结束时间"
                  },
                  "content": {
                    "type": "string",
                    "description": "上课内容"
                  },
                  "homework": {
                    "type": "string",
                    "description": "作业"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse"
                }
              }
            }
          }
        }
      }
    },
    "/deduction/manual": {
      "post": {
        "summary": "手动扣课",
        "description": "手动为学生扣减课时",
        "tags": ["扣课管理"],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "studentId": {
                    "type": "integer",
                    "description": "学生ID"
                  },
                  "courseId": {
                    "type": "integer",
                    "description": "课程ID"
                  },
                  "deductionAmount": {
                    "type": "number",
                    "description": "扣课数量"
                  },
                  "remark": {
                    "type": "string",
                    "description": "备注"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "扣课成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse"
                }
              }
            }
          }
        }
      }
    }
  }
}